+++ tests/test_cli_workflows.py
@@ -50,14 +50,14 @@ def test_workflows_server_ok():
         "items": [
             {
                 "status": "running",
-                "created": "2018-06-13T09:47:35.66097",
+                "created": "2018-06-13T09:47:35",
                 "user": "00000000-0000-0000-0000-000000000000",
                 "name": "mytest.1",
                 "id": "256b25f4-4cfb-4684-b7a8-73872ef455a1",
                 "size": {"raw": 0, "human_readable": "0 Bytes"},
                 "progress": {
-                    "run_started_at": "2018-06-13T09:47:40.28223",
-                    "run_finished_at": "2018-06-13T10:30:03.70303",
+                    "run_started_at": "2018-06-13T09:47:40",
+                    "run_finished_at": "2018-06-13T10:30:03",
                 },
             }
         ]
@@ -120,19 +120,19 @@ def test_workflows_sorting():
         "items": [
             {
                 "status": "running",
-                "created": "2018-06-13T09:47:35.66097",
+                "created": "2018-06-13T09:47:35",
                 "user": "00000000-0000-0000-0000-000000000000",
                 "name": "mytest.1",
                 "id": "256b25f4-4cfb-4684-b7a8-73872ef455a1",
                 "progress": {
-                    "run_started_at": "2018-06-13T09:47:40.28223",
-                    "run_finished_at": "2018-06-13T10:30:03.70303",
+                    "run_started_at": "2018-06-13T09:47:40",
+                    "run_finished_at": "2018-06-13T10:30:03",
                 },
                 "size": {"raw": 0, "human_readable": "0 Bytes"},
             },
             {
                 "status": "running",
-                "created": "2018-06-13T09:55:35.66097",
+                "created": "2018-06-13T09:55:35",
                 "user": "00000000-0000-0000-0000-000000000000",
                 "name": "mytest.2",
                 "id": "256b25f4-4cfb-4684-b7a8-73872ef455a2",
@@ -156,12 +156,8 @@ def test_workflows_sorting():
                 cli, ["list", "-t", reana_token, "--sort", "run_number"]
             )
             message = (
-                "mytest   2            2018-06-13T09:55:35.66097   -"
-                "                           -"
-                "                           running\n"
-                "mytest   1            2018-06-13T09:47:35.66097   "
-                "2018-06-13T09:47:40.28223   2018-06-13T10:30:03.70303"
-                "   running"
+                "mytest   2            2018-06-13T09:55:35   -                     -                     running\n"
+                "mytest   1            2018-06-13T09:47:35   2018-06-13T09:47:40   2018-06-13T10:30:03   running"
             )
             assert result.exit_code == 0
             assert message in result.output